requirements.txt
discord.py[voice]>=2.7.1
davey>=0.1.4
yt-dlp>=2024.12.0
pynacl>=1.5.0
spotipy>=2.24.0
redis>=5.0.0
